Move API endpoints to api.js
